refactor: rename json-encoding.ts to json-encoding-modern.ts
Rename source and test files, update import in json-encoding-dispatch.ts,
and update deno.json export entry. No cross-package imports of the old
path existed (only json-encoding-dispatch is imported externally).
