+++ packages/data-model/json-encoding-modern.ts
